+++ src/flake8_custom_import_rules/utils/parse_utils.py
@@ -34,17 +34,49 @@ def parse_module_string(
     suffix: str | None = None,
     delimiter: str = ".",
 ) -> list:
-    """Parse a module."""
+    """Parse a module and return substrings that match certain criteria.
+
+    Parameters
+    ----------
+    value : str
+        The string to parse.
+    substring_match : list | str | None
+        A list of substrings or a single substring to match against. If a
+        substring in the value matches, it will be returned in the result.
+    prefix : str | None
+        A prefix to match against. If a substring in the value has this prefix,
+        it will be returned in the result.
+    suffix : str | None
+        A suffix to match against. If a substring in the value has this suffix,
+        it will be returned in the result.
+    delimiter : str
+        The delimiter to use when splitting the value.
+
+    Returns
+    -------
+    list
+        A list of substrings from the value that match the provided criteria.
+    """
     substrings = value.split(delimiter)
     if isinstance(substring_match, str):
         substring_match = [substring_match]
+
     if substring_match or prefix or suffix:
+        checking_private_identifiers = prefix == "_"
         matches: list[str] = []
         matches.extend(
             substring
             for substring in substrings
             if (substring_match and substring in substring_match)
-            or (prefix and substring.startswith(prefix))
+            or (
+                prefix
+                and substring.startswith(prefix)
+                and not (
+                    checking_private_identifiers
+                    and substring.startswith("__")
+                    and substring.endswith("__")
+                )
+            )
             or (suffix and substring.endswith(suffix))
         )
         return matches
@@ -59,7 +91,30 @@ def check_string(
     suffix: str | None = None,
     delimiter: str = ".",
 ) -> bool:
-    """Check string for matches."""
+    """Check a string or list of strings for matches against certain criteria.
+
+    Parameters
+    ----------
+    strings_to_check : list | str
+        The string(s) to check.
+    substring_match : list | str | None
+        A list of substrings or a single substring to match against. If a
+        substring in the string(s) to check matches, the function will return
+        True.
+    prefix : str | None
+        A prefix to match against. If a string in the string(s) to check has
+        this prefix, the function will return True.
+    suffix : str | None
+        A suffix to match against. If a string in the string(s) to check has
+        this suffix, the function will return True.
+    delimiter : str
+        The delimiter to use when splitting the string(s) to check.
+
+    Returns
+    -------
+    bool
+        True if any matches are found, False otherwise.
+    """
     if substring_match or prefix or suffix:
         if isinstance(strings_to_check, list):
             strings_to_check = f"{delimiter}".join(strings_to_check)